requirements.txt
absl-py==2.1.0
ase==3.22.1
astunparse==1.6.3
botorch==0.9.2
cachetools==5.3.3
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.4.1
contourpy==1.1.0
copulas==0.9.2
cycler==0.11.0
Cython==3.0.2
decorator==5.1.1
emmet-core==0.68.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.1
future==0.18.3
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
GPUtil==1.4.0
GPy==1.10.0
GPyOpt==1.2.6
gpytorch==1.11
grpcio==1.62.0
h5py==3.10.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.0.1
jaxtyping==0.2.22
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
latexcodec==2.0.1
libclang==16.0.6
linear-operator==0.5.1
lit==16.0.6
llvmlite==0.42.0
m3gnet==0.2.4
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.3
ml-dtypes==0.2.0
monty==2023.9.5
mp-api==0.36.1
mpmath==1.3.0
msgpack==1.0.5
multipledispatch==1.0.0
networkx==3.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
palettable==3.3.3
pandas==1.2.0
paramz==0.9.5
Pillow==10.0.0
plotly==5.16.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybtex==0.24.0
pydantic==1.10.12
pymatgen==2023.9.10
pynndescent==0.5.12
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.8.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.13.0
six==1.16.0
spglib==2.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.0.1
tqdm==4.66.1
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.6
uncertainties==3.1.7
urllib3==2.0.4
Werkzeug==3.0.1
wrapt==1.14.1
xgboost==2.0.0
zipp==3.16.2
